requirements.txt
pandas
python-dotenv
langchain
langgraph
langchain-community
azure-identity
azure-storage-blob
azure-data-tables
langchain-openai
sqlalchemy
pyodbc
azure-monitor-opentelemetry-exporter
fastapi
uvicorn
azure-functions
